git-svn-id: https://svn.osdn.net/svnroot/switch/switch/trunk@624 6544d933-5e26-41f1-b2f0-1bd7b5b7622e

+++ src/svvitch/MediaItem.h
@@ -106,7 +106,7 @@ public:
 	const DWORD getHexProperty(const string& key, const DWORD& defaultValue) const {
 		Poco::UInt64 num = defaultValue;
 		Poco::NumberParser::tryParseHex64(getProperty(key), num);
-		return num;
+		return (DWORD)num;
 	}
 
 	const double getFloatProperty(const string& key, const double& defaultValue) const {
